# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Nutikas Restorani Reserveerimissusteem API
  version: 1.0.0
  description: >
    Spring Boot API contract for the floorplan-first restaurant demo with
    plan tabs, deterministic demo occupancy, and table recommendations.
servers:
  - url: /api/v1
    description: API version 1 base path

paths:
  /layout:
    get:
      summary: Get floorplan layout
      tags:
        - layout
      operationId: getLayout
      responses:
        "200":
          description: Floorplan geometry, plans, zones, and visual features
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LayoutResponse"

  /availability:
    post:
      summary: Get table availability for a plan, time, and party
      tags:
        - availability
      operationId: getAvailability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilityRequest"
      responses:
        "200":
          description: Availability state for tables in the selected plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /availability/slots:
    post:
      summary: Get nearby available times for the selected plan and party
      tags:
        - availability
      operationId: getAvailabilitySlots
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilitySlotsRequest"
      responses:
        "200":
          description: Nearby times with at least one available table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilitySlotsResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /recommendations:
    post:
      summary: Get ordered recommendations for available tables
      tags:
        - recommendations
      operationId: getRecommendations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendationRequest"
      responses:
        "200":
          description: Ordered recommendations and top recommendation id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationsResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reservations:
    post:
      summary: Confirm reservation on selected table
      tags:
        - reservations
      operationId: createReservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationRequest"
      responses:
        "201":
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationResponse"
        "409":
          description: Selected table already reserved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reservations/{id}:
    get:
      summary: Get reservation details
      tags:
        - reservations
      operationId: getReservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reservation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationResponse"
        "404":
          description: Reservation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    Zone:
      type: string
      enum:
        - INDOOR
        - TERRACE
        - PRIVATE

    PlanCode:
      type: string
      enum:
        - INDOOR
        - TERRACE

    FeatureType:
      type: string
      enum:
        - PRIVATE_ROOM
        - PLAY_AREA
        - WINDOW_BAND

    AvailabilityStatus:
      type: string
      description: >
        Availability endpoint returns only booking-availability states.
      enum:
        - AVAILABLE
        - RESERVED
        - UNAVAILABLE

    Preference:
      type: string
      enum:
        - PRIVACY
        - WINDOW
        - ACCESSIBLE
        - NEAR_PLAY_AREA

    RecommendationPreference:
      type: string
      enum:
        - PRIVACY
        - WINDOW
        - NEAR_PLAY_AREA

    ReservationStatus:
      type: string
      enum:
        - CONFIRMED
        - REJECTED
        - CANCELLED

    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        details:
          type: array
          items:
            type: string
      required:
        - status
        - message

    Point:
      type: object
      properties:
        x:
          type: number
        y:
          type: number
      required:
        - x
        - y

    PlanSummary:
      type: object
      properties:
        code:
          $ref: "#/components/schemas/PlanCode"
        label:
          type: string
        description:
          type: string
      required:
        - code
        - label

    ZoneSummary:
      type: object
      properties:
        code:
          $ref: "#/components/schemas/Zone"
        label:
          type: string
        description:
          type: string
      required:
        - code
        - label

    LayoutFeature:
      type: object
      properties:
        featureId:
          type: string
        plan:
          $ref: "#/components/schemas/PlanCode"
        type:
          $ref: "#/components/schemas/FeatureType"
        label:
          type: string
        center:
          $ref: "#/components/schemas/Point"
        width:
          type: number
        height:
          type: number
      required:
        - featureId
        - plan
        - type
        - label
        - center
        - width
        - height

    TableGeometry:
      type: object
      properties:
        tableId:
          type: string
        label:
          type: string
        capacity:
          type: integer
          minimum: 1
        plan:
          $ref: "#/components/schemas/PlanCode"
        zone:
          $ref: "#/components/schemas/Zone"
        center:
          $ref: "#/components/schemas/Point"
        width:
          type: number
        height:
          type: number
        rotationDegrees:
          type: number
        accessible:
          type: boolean
        nearWindow:
          type: boolean
          nullable: true
        nearPlayArea:
          type: boolean
          nullable: true
        privacyScore:
          type: number
          minimum: 0
          maximum: 1
      required:
        - tableId
        - label
        - capacity
        - plan
        - zone
        - center
        - width
        - height
        - rotationDegrees
        - accessible
        - nearWindow
        - nearPlayArea
        - privacyScore

    LayoutResponse:
      type: object
      properties:
        layoutId:
          type: string
        venueWidthMeters:
          type: number
        venueHeightMeters:
          type: number
        plans:
          type: array
          items:
            $ref: "#/components/schemas/PlanSummary"
        zones:
          type: array
          items:
            $ref: "#/components/schemas/ZoneSummary"
        features:
          type: array
          items:
            $ref: "#/components/schemas/LayoutFeature"
        tables:
          type: array
          items:
            $ref: "#/components/schemas/TableGeometry"
      required:
        - layoutId
        - venueWidthMeters
        - venueHeightMeters
        - plans
        - zones
        - features
        - tables

    AvailabilityRequest:
      type: object
      properties:
        date:
          type: string
          format: date
        time:
          type: string
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
        partySize:
          type: integer
          minimum: 1
        plan:
          $ref: "#/components/schemas/PlanCode"
        zone:
          allOf:
            - $ref: "#/components/schemas/Zone"
          description: >
            Optional filter. If provided, non-matching tables inside the selected plan
            are marked UNAVAILABLE.
        accessibleRequired:
          type: boolean
          description: >
            Hard requirement. If true, non-accessible tables inside the selected plan
            are marked UNAVAILABLE.
      required:
        - date
        - time
        - partySize
        - plan
        - accessibleRequired

    RecommendationRequest:
      type: object
      allOf:
        - $ref: "#/components/schemas/AvailabilityRequest"
        - type: object
          properties:
            preferences:
              type: array
              items:
                $ref: "#/components/schemas/RecommendationPreference"
              uniqueItems: true
              description: >
                Used to calculate recommendation score and explanation reasons.

    AvailabilityResponse:
      type: object
      properties:
        tableStatusById:
          description: >
            Map keyed by tableId for tables in the selected plan only.
          type: object
          additionalProperties:
            $ref: "#/components/schemas/AvailabilityStatus"
        generatedAt:
          type: string
          format: date-time
      required:
        - tableStatusById
        - generatedAt

    AvailabilitySlotsRequest:
      type: object
      allOf:
        - $ref: "#/components/schemas/AvailabilityRequest"
        - type: object
          properties:
            preferences:
              type: array
              items:
                $ref: "#/components/schemas/RecommendationPreference"
              uniqueItems: true
              description: >
                Used only to determine each slot's topRecommendationId.

    AvailabilitySlot:
      type: object
      properties:
        time:
          type: string
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
        availableTableCount:
          type: integer
          minimum: 0
        topRecommendationId:
          type: string
          nullable: true
      required:
        - time
        - availableTableCount
        - topRecommendationId

    AvailabilitySlotsResponse:
      type: object
      properties:
        requestedTime:
          type: string
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
        slots:
          type: array
          items:
            $ref: "#/components/schemas/AvailabilitySlot"
        generatedAt:
          type: string
          format: date-time
      required:
        - requestedTime
        - slots
        - generatedAt

    Recommendation:
      type: object
      properties:
        tableId:
          type: string
        score:
          type: number
          minimum: 0
          maximum: 100
        reasons:
          type: array
          items:
            type: string
      required:
        - tableId
        - score
        - reasons

    RecommendationsResponse:
      type: object
      properties:
        topRecommendationId:
          type: string
          nullable: true
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/Recommendation"
      required:
        - topRecommendationId
        - recommendations

    ReservationRequest:
      type: object
      properties:
        date:
          type: string
          format: date
        time:
          type: string
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
        partySize:
          type: integer
          minimum: 1
        tableId:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: string
          format: email
        preferences:
          type: array
          items:
            $ref: "#/components/schemas/Preference"
      required:
        - date
        - time
        - partySize
        - tableId

    ReservationResponse:
      type: object
      properties:
        reservationId:
          type: string
        tableId:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
        partySize:
          type: integer
          minimum: 1
        status:
          $ref: "#/components/schemas/ReservationStatus"
        createdAt:
          type: string
          format: date-time
        message:
          type: string
      required:
        - reservationId
        - tableId
        - date
        - time
        - partySize
        - status
        - createdAt